+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - Purge command
  - Superuser shutdown and restart commands
  - Superuser check to admin commands
+ - Setguild command
 ### Changed
  - Readme's TODO updated
  - Balance command to bal